Test/requirements.txt
Flask==2.3.3
python-gpiod>=1.5.4; platform_system!='Windows'